Support iPad
Check if device is iPad. If so, render pdf as usual. pluginFound now returns "iPad".

Also remove default initialisation of width and height. If they are not set, they are either way initialised on line 225,226. If device is iPad, and user has not set width and height, use "huge" values. iPad flattens the rendering, i.e. the pdf content is not "scrollable" so everything has to be visible.