Reworked logic to support navigator.pdfViewerEnabled
The logic is spelled out in code comments, but the gist is that we only execute tests when needed, not all up front as previously written. We also add a check for navigator.pdfViewerEnabled prior to doing any sniffing (we can avoid sniffing if navigator.pdfViewerEnabled is available and true). Finally, we revert to prior logic involving assumptions about modern browsers having native PDF support, and IE11 using ActiveX.

These changes modernize the approach, as should be more performant as it reduces the number of browser checks